transform to BigDecimal for calculation

+++ app/src/main/java/de/ingoreschke/increasesalarycalculator/SalaryIncreasePresenter.kt
@@ -1,8 +1,10 @@
 package de.ingoreschke.increasesalarycalculator
 
+import java.math.BigDecimal
+
 class SalaryIncreasePresenter(private val salaryIncreaseInteractor: SalaryIncreaseInteractor) {
 
-    fun calculateSalaryIncrease(salary: Double, increasePercentage: Double): Double {
+    fun calculateSalaryIncrease(salary: BigDecimal, increasePercentage: BigDecimal): BigDecimal {
         return salaryIncreaseInteractor.calcIncreasedSalary(salary, increasePercentage)
     }
 }
